refactor: /app 에서의 불필요한 import graph 문제 해결 (번들 726)
