+++ apps/web/src/features/notifications/query.ts
@@ -1,0 +1,1 @@
+export { useUnreadNotificationsCountQuery } from './api/notifications-query';
