+++ apps/web/src/features/stretching-session/query.ts
@@ -1,0 +1,5 @@
+export type {
+  ValidStretchingSessionsQueryKey,
+  ValidStretchingSessionsQueryOptionsType,
+} from './api/useValidStretchingSessionsQuery';
+export { useValidStretchingSessionsQuery } from './api/useValidStretchingSessionsQuery';
