+++ apps/web/src/features/user-profile/query.ts
@@ -1,0 +1,2 @@
+export type { UserCharacter, UserMeData, UserMeResponse, UserProfile } from './api/types';
+export { useUserProfileQuery } from './api/user-profile-query';
