+++ apps/web/src/widgets/user-status-card/ui/UserStatusCardSection.tsx
@@ -1,6 +1,6 @@
 import { UserStatusCard } from '@repo/ui/user-status-card';
 
-import { useUserProfileQuery } from '@/src/features/user-profile';
+import { useUserProfileQuery } from '@/src/features/user-profile/query';
 import StreakImoji from '@/src/shared/assets/streak-imoji.svg';
 import { buildImageUrl } from '@/src/shared/lib/image';
